requirements.txt
apistar==0.1.17
bibtexparser==0.6.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
condor-ir==1.1.0
coreapi==2.3.1
coreschema==0.0.4
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==2.13.5
multidict==2.1.5
nltk==3.2.3
numpy==1.12.1
py==1.4.34
pyenchant==1.6.8
PyPDF2==1.26.0
pytest==3.1.2
requests==2.18.1
six==1.10.0
SQLAlchemy==1.0.12
tabulate==0.7.7
tqdm==4.11.2
uritemplate==3.0.0
urllib3==1.21.1
Werkzeug==0.12.2
whitenoise==3.3.0
yarl==0.10.2
coverage==4.4.1
